% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnel.R
\name{BrdQtrlyFunnel}
\alias{BrdQtrlyFunnel}
\title{Brand Quarterly Funnel}
\usage{
BrdQtrlyFunnel()
}
\description{
Calculate Familiarity (top 5 box / all respondents answered Fam), 
   FO (top 3 box / all respondents answered Opinion), Consideration (top 2 box), 
   Shop (shopped certain brand / total shoppers), Intention (intend / total respondents).
}

