% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funnel.R
\name{Input}
\alias{Input}
\title{Input Function}
\usage{
Input(vname)
}
\arguments{
\item{vname}{Character, such as "market", "brand", "quarter"...}
}
\value{
Character vector, specified values of the "vname".
}
\description{
User friendly system for inputing parameters of other functions
}

